requirements.txt
PySide6==6.5.2
cryptography==41.0.3
selenium==4.12.0
webdriver-manager==4.0.0
schedule==1.2.0
pillow==11.1.0 